@@ -1,28 +1,32 @@
 {
-  "_from": "body-parser@1.18.3",
-  "_id": "body-parser@1.18.3",
+  "_from": "body-parser",
+  "_id": "body-parser@1.19.0",
   "_inBundle": false,
-  "_integrity": "sha1-WykhmP/dVTs6DyDe0FkrlWlVyLQ=",
+  "_integrity": "sha512-dhEPs72UPbDnAQJ9ZKMNTP6ptJaionhP5cBb541nXPlW60Jepo9RV/a4fX4XWW9CuFNK22krhrj1+rgzifNCsw==",
   "_location": "/body-parser",
-  "_phantomChildren": {},
+  "_phantomChildren": {
+    "depd": "1.1.2",
+    "inherits": "2.0.3",
+    "toidentifier": "1.0.0"
+  },
   "_requested": {
-    "type": "version",
+    "type": "tag",
     "registry": true,
-    "raw": "body-parser@1.18.3",
+    "raw": "body-parser",
     "name": "body-parser",
     "escapedName": "body-parser",
-    "rawSpec": "1.18.3",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "1.18.3"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/express",
-    "/karma"
+    "#USER",
+    "/"
   ],
-  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.18.3.tgz",
-  "_shasum": "5b292198ffdd553b3a0f20ded0592b956955c8b4",
-  "_spec": "body-parser@1.18.3",
-  "_where": "D:\\Projects\\ITI\\Angular\\Angular-Project\\e-commerce\\node_modules\\express",
+  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.19.0.tgz",
+  "_shasum": "96b2709e57c9c4e09a6fd66a8fd979844f69f08a",
+  "_spec": "body-parser",
+  "_where": "C:\\Users\\Basma\\Downloads\\e-commerce-mean-stack\\NodeServer",
   "bugs": {
     "url": "https://github.com/expressjs/body-parser/issues"
   },
@@ -39,32 +43,32 @@
     }
   ],
   "dependencies": {
-    "bytes": "3.0.0",
+    "bytes": "3.1.0",
     "content-type": "~1.0.4",
     "debug": "2.6.9",
     "depd": "~1.1.2",
-    "http-errors": "~1.6.3",
-    "iconv-lite": "0.4.23",
+    "http-errors": "1.7.2",
+    "iconv-lite": "0.4.24",
     "on-finished": "~2.3.0",
-    "qs": "6.5.2",
-    "raw-body": "2.3.3",
-    "type-is": "~1.6.16"
+    "qs": "6.7.0",
+    "raw-body": "2.4.0",
+    "type-is": "~1.6.17"
   },
   "deprecated": false,
   "description": "Node.js body parsing middleware",
   "devDependencies": {
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.11.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.7.0",
-    "eslint-plugin-standard": "3.1.0",
+    "eslint": "5.16.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.17.2",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-promise": "4.1.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
     "methods": "1.1.2",
-    "mocha": "2.5.3",
+    "mocha": "6.1.4",
     "safe-buffer": "5.1.2",
-    "supertest": "1.1.0"
+    "supertest": "4.0.2"
   },
   "engines": {
     "node": ">= 0.8"
@@ -88,5 +92,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --require test/support/env --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --require test/support/env --reporter spec --check-leaks test/"
   },
-  "version": "1.18.3"
+  "version": "1.19.0"
 }
